+++ src/main/kotlin/net/ndrei/teslacorelib/render/selfrendering/IDrawable.kt
@@ -1,0 +1,8 @@
+package net.ndrei.teslacorelib.render.selfrendering
+
+import net.minecraft.client.renderer.BufferBuilder
+
+@FunctionalInterface
+interface IDrawable {
+    fun draw(buffer: BufferBuilder)
+}